+++ tests/testthat/test-er_opi.R
@@ -47,7 +47,7 @@ test_that("check er_opi", {
 test_that("check er_opi", {
   expect_equal(
     test$dt.farm.score,
-    expected = 66,
+    expected = 77,
     tolerance = 0.01)
 })